Remove DrupalProjectTrait
